+++ src/plugin-store-item.svelte
@@ -51,7 +51,7 @@ function installExternalPlugin(plugin: Plugin) {
 	const currentPlugins = getStoredPlugins();
 
 	const pluginCopy = { ...plugin };
-	pluginCopy.installed = true;
+	pluginCopy.active = true;
 	currentPlugins.push(pluginCopy);
 
 	localPlugins = currentPlugins; 
@@ -64,22 +64,23 @@ function installExternalPlugin(plugin: Plugin) {
 function uninstallExternalPlugin(plugin: Plugin) {
 	const currentPlugins = getStoredPlugins();
 	const updatedPlugins = currentPlugins.filter((it) => it.name !== plugin.name);
+	
 	localPlugins = updatedPlugins;
 
 	dispatchConfigurePlugin(plugin, true);
 	console.log("Uninstalled external plugin:", plugin.name);
 }
 
-// Enables/disables plugin by toggling the "installed" property.
+// Enables/disables plugin by toggling the "active" property.
 function toggleOfficialPlugin(plugin: Plugin, isEnabled: boolean) {
 	const currentPlugins = getStoredPlugins();
 	const foundPlugin = currentPlugins.find((it) => it.name === plugin.name);
-	if (foundPlugin) {
-		foundPlugin.installed = isEnabled;
-	}
-
+    if (foundPlugin) {
+        foundPlugin.active = isEnabled;
+    }
+    
 	localPlugins = currentPlugins;
-	plugin.installed = isEnabled;
+    plugin.active = isEnabled;
 
 	dispatchConfigurePlugin(plugin);
 	console.log("Set toggle state for", plugin.name);
@@ -150,7 +151,7 @@ function getPluginIcon(plugin: Plugin) {
             </div>
         </plugin-store-item-meta--wrapper>
     </plugin-store-item-meta>
-    {#if plugin.installed}
+    {#if plugin.active}
         {#if plugin.official}
             <Button variant="outlined" onclick={() => toggleOfficialPlugin(plugin, false)}>
                 Disable
@@ -171,7 +172,7 @@ function getPluginIcon(plugin: Plugin) {
             Enable
         </Button>
     {:else}
-        {#if localPlugins.includes(plugin)}
+        {#if localPlugins.some(p => p.name === plugin.name)}
             <SplitButton label="Enable" onclick={() => {toggleOfficialPlugin(plugin, true)}} onmenuOpen={() => openPluginMenu(index)}>
                 <Menu bind:this={menus[index]} open={menuStates[index]} anchorCorner="BOTTOM_LEFT" style="left: -70px;">
                     <List>