+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Migrated to Svelte 5
+- `installed` is not a property of Plugin anymore
 
 ### Removed
 - Removed customElement as it isn't needed